Discussion fix
